% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_attrs.R
\name{add_edge_color}
\alias{add_edge_color}
\title{Modify the edge colors of a sigmagraph object.}
\usage{
add_edge_color(
  sigma_obj,
  one_color = NULL,
  color_attr = NULL,
  color_palette = "Set2"
)
}
\arguments{
\item{sigma_obj}{Sigmagraph object}

\item{one_color}{A single color to color all of the nodes (hex format)}
}
\value{
Sigmagraph with modified edge colors
}
\description{
Modify the edge colors of a sigmagraph object by providing a single color.
Also works with a vector of correct size.
}
\examples{
library(igraph)
library(sigmagraph)

data(lesMis)

sig <- sigma_from_igraph(graph = lesMis) \%>\%
  add_edge_color(one_color = "#ccc")
sig

}
