% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_attrs.R
\name{add_edge_size}
\alias{add_edge_size}
\title{Modify the edge size of a sigmagraph object.}
\usage{
add_edge_size(sigma_obj, one_size = NULL)
}
\arguments{
\item{sigma_obj}{Sigmagraph object}

\item{one_size}{A single size to use for all edges}
}
\value{
Sigmagraph with modified edge sizes
}
\description{
Modify the edge size of a sigmagraph by providing a single size
}
\examples{
library(igraph)
library(sigmagraph)

data(lesMis)

sig <- sigma_from_igraph(graph = lesMis) \%>\%
  add_edge_size(one_size = 0.1)
sig

}
