% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_attrs.R
\name{add_edge_zindex}
\alias{add_edge_zindex}
\title{Modify the edge zIndex of a sigmagraph object.}
\usage{
add_edge_zindex(sigma_obj, zindex)
}
\arguments{
\item{sigma_obj}{Sigmagraph object}

\item{zindex}{Zindex value, larger is drawn above.}
}
\value{
Sigmagraph
}
\description{
Modify the edge zIndex
}
\examples{
library(igraph)
library(sigmagraph)

data(lesMis)

sig <- sigma_from_igraph(graph = lesMis) \%>\%
  add_edge_zindex(zindex = 2)
sig

}
