% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_attrs.R
\name{add_node_labels}
\alias{add_node_labels}
\title{Modify the node labels of a sigmagraph object.}
\usage{
add_node_labels(sigma_obj, label_attr = NULL)
}
\arguments{
\item{sigma_obj}{Sigmagraph object, returned by sigma_from_igraph function}

\item{label_attr}{Attribute to use to replace node labels}
}
\value{
A sigmagraph object with modified node labels
}
\description{
Modify the node labels of an existing sigmagraph object by providing an
attribute from the initial igraph to use as the labels.
}
\examples{
library(igraph)
library(sigmagraph)
library(magrittr)

data(lesMis)

layout <- layout_nicely(lesMis)
sig <- sigma_from_igraph(graph = lesMis, layout = layout) \%>\%
  add_node_labels(label_attr = 'label')
sig

}
