% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_shiny.R
\name{renderSigmagraph}
\alias{renderSigmagraph}
\title{Render a sigmagraph visualization in Shiny}
\usage{
renderSigmagraph(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that creates a sigmagraph visualization}

\item{env}{Defaults to parent.frame() (cf. Shiny docs)}

\item{quoted}{Defaults to FALSE (cf. Shiny docs)}
}
\description{
Render a sigmagraph visualization in Shiny
}
