% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_shiny.R
\name{sigma_from_igraph}
\alias{sigma_from_igraph}
\title{Build a sigmagraph object from an igraph object}
\usage{
sigma_from_igraph(
  graph,
  layout = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{graph}{Igraph object}

\item{layout}{Output of an igraph layout function (default: layout_nicely)}

\item{width}{Width of the output graph (default: fit container)}

\item{height}{Height of the output graph (default: fit container)}

\item{elementId}{Do not specify, used by the htmlwidgets package}
}
\value{
Htmlwidget object, meant to be called directly to render a default
  visualization, or passed to other functions to change attributes
  (colors, sizes, interactivity, etc.).
}
\description{
Basic sigma.js visualization of an igraph object, with pipeable syntax.
}
\examples{
library(igraph)
library(sigmagraph)

data(lesMis)

l <- layout_nicely(lesMis)
sig <- sigma_from_igraph(graph = lesMis, layout = l)

sig
}
