% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_shiny.R
\name{sigmagraphOutput}
\alias{sigmagraphOutput}
\title{Create a UI element for a sigmagraph visualization in Shiny}
\usage{
sigmagraphOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{ID of the UI element}

\item{width}{Width of the UI element}

\item{height}{Height of the UI element}
}
\description{
Create a UI element for a sigmagraph visualization in Shiny
}
